Внесение изменений в задание 32
